server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.9.0
bcrypt==4.2.0
pydantic==2.9.2
email-validator==2.2.0
python-multipart==0.0.10
